% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gData1}
\alias{gData1}
\title{gData1 (Gaussian Dataset 1)}
\format{
\subsection{}{

A data frame with 400 rows and 3 dimenions. First two ones are (x,y) coordinates, followed by class labels.
}
}
\source{
Jaskowiak, P.A., Costa, I.G. & Campello, R.J.G.B. The area under the ROC curve as a measure of clustering quality. Data Min Knowl Disc 36, 1219–1245 (2022). \url{https://doi.org/10.1007/s10618-022-00829-0}.
}
\usage{
gData1
}
\description{
Gaussian Dataset 1. A 2D dataset with two Gaussian (globular) clusters.
}
\details{
Clusters are centered at (0,0) and (40,40), with same variance (25) for box clusters and axis.

This dataset appears in Fig. 5(a) of the referenced paper.
}
\keyword{datasets}
