% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ruspini}
\alias{ruspini}
\title{Ruspini Data}
\format{
\subsection{}{

A data frame with 75 rows and 2 dimenions (x,y), no true class labels
}
}
\source{
E. H. Ruspini (1970) Numerical methods for fuzzy clustering. Inform. Sci. 2, 319–350.
}
\usage{
ruspini
}
\description{
The Ruspini dataset. A 2D dataset usually employed to exemplify clustering
methods and/or validation techniques.
}
\keyword{datasets}
