% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucc.R
\name{aucc}
\alias{aucc}
\title{aucc}
\usage{
aucc(partition, dataset = NULL, distance = NULL, distanceMethod = "euclidean")
}
\arguments{
\item{partition}{A hard/crisp partition}

\item{dataset}{The dataset from which the partition was obtained}

\item{distance}{Distances between object pairs, obtained with the dist function}

\item{distanceMethod}{The distance method employed to compute distances from the dataset (see ?dist)}
}
\value{
The AUCC of the partition (clustering solution)
}
\description{
Computes the Area Under the (ROC) Curve for Clustering (AUCC)
}
\details{
Function computes the Area Under the (ROC) Curve for Clustering (AUCC) of a clustering solution.

You need to provide a partitioning of the data (clustering solution) and one of the following:
(i)   the dataset from which the partition is originated;
(ii)  the distance matrix from which the partition was originated.

In case of (i), you may also specify the distance function that will be employed to
compute the distances internally by aucc, using the distanceMethod parameter. If you do not
specify any string, it defaults to the Euclidean Distance. Make sure to pass strings accepted
by the function dist (package stats).

In case of (ii), make sure that the distance matrix was originated/converted with the function
dist from the package stats, using the parameters diag = FALSE, upper = FALSE (these are the
defaults for dist). Other than that, you may choose any distance / dissimilarity you find
appropriate for your data. If you have to perform multiple computations of aucc for the same
dataset, with different partitions, prefer this approach as computing the distance matrix
is computationally expensive.

Make sure to validate your partition with the same distance you used to generate it.
}
\examples{

\dontrun{
library(cluster) #for ruspini dataset and k-means

# Computing AUCC from the dataset itself

clustering <- kmeans(ruspini,2)$cluster
aucc(clustering, ruspini)

# Computing AUCC from a precomputed distance matrix and selecting best k
myDistance <- dist(ruspini)

auccValues <- c()

for (k in 2:ceiling(sqrt(dim(ruspini)[1]))) {
   clustering <- kmeans(ruspini,k)$cluster
   auccValues <- append(auccValues, aucc(clustering, distance = myDistance))
}

sprintf('Suggested number of clusters is: \%d', which.max(auccValues)+1)

}

}
\references{
Jaskowiak, P.A., Costa, I.G. & Campello, R.J.G.B. The area under the ROC curve as a measure of clustering quality. Data Mining and Knowledge Discovery 36, 1219–1245 (2022). https://doi.org/10.1007/s10618-022-00829-0
}
